refactor folders, imports, components. Fix: login/registerPage
