+++ src/app/HOC/ProtectedRoute/index.tsx
@@ -1,5 +1,6 @@
 import { FC } from 'react'
 import { Navigate } from 'react-router-dom'
+
 import { useAuth } from '../../hooks/use-auth'
 
 export const ProtectedRoute: FC<any> = ({ component: Component, ...props }) => {
