+++ src/app/api/index.ts
@@ -1,0 +1,1 @@
+export * as UnsplashApi from './unsplash'
