+++ src/components/Card/index.tsx
@@ -1,12 +1,13 @@
-import React from 'react'
+import React, { FC } from 'react'
 import { Link } from 'react-router-dom'
 
 import styles from './Card.module.scss'
 
-import fav from '../../../assets/icons/fav.png'
-import favFilled from '../../../assets/icons/fav-filled.png'
+import favFilled from '../../assets/icons/fav-filled.png'
+import fav from '../../assets/icons/fav.png'
 
-function Card({ id, urls, description = 'title', liked_by_user }) {
+// TODO: Катя, доделай пожалуйста propsType
+const Card: FC<any> = ({ id, urls, description = 'title', liked_by_user }) => {
   const [isLiked, setIsLiked] = React.useState(liked_by_user)
   const handleLikePost = () => {
     setIsLiked(() => !isLiked)
