+++ src/components/Preloader/Preloader.tsx
@@ -1,0 +1,5 @@
+const Preloader = () => {
+  return <img src='' alt='Preloader' />
+}
+
+export default Preloader
