--- src/components/organisms/index.ts
@@ -1,3 +1,0 @@
-export { default as Card } from './Card'
-export * from './CardList'
-export * from './Search'
