+++ src/index.scss
@@ -1,5 +1,5 @@
-@import url('./vendor/normalize.css');
-@import url('./vendor/fonts/inter/fonts.css');
+@import url('./app/vendor/normalize.css');
+@import url('./app/vendor/fonts/inter/fonts.css');
 
 :root {
   --max-width: 1280px;
