+++ src/index.tsx
@@ -2,10 +2,11 @@ import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import { Provider } from 'react-redux'
 import { BrowserRouter } from 'react-router-dom'
+
 import App from './App'
+import { CURRENT_BASE_URL } from './app/constants/environment'
 import './index.scss'
 import { store } from './store'
-import { CURRENT_BASE_URL } from './constants/enviroment'
 
 const root = createRoot(document.getElementById('root') as HTMLElement)
