--- src/pages/LoginPage.tsx
@@ -1,16 +1,0 @@
-import { Link } from 'react-router-dom'
-import { Login } from '../components/LoginForm/Login'
-
-const LoginPage = () => {
-  return (
-    <div>
-      <h1>Login</h1>
-      <Login />
-      <p>
-        Or <Link to='/register'>register</Link>
-      </p>
-    </div>
-  )
-}
-
-export default LoginPage
