--- src/pages/RegisterPage.tsx
@@ -1,16 +1,0 @@
-import { Link } from 'react-router-dom'
-import { SignUp } from '../components/LoginForm/SignUp'
-
-const RegisterPage = () => {
-  return (
-    <div>
-      <h1>Register</h1>
-      <SignUp />
-      <p>
-        Already have an account? <Link to='/login'>Sign in</Link>
-      </p>
-    </div>
-  )
-}
-
-export default RegisterPage
